Define hand ordering strategy inside GameCore
